pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cybercar"
version = "0.1.0"
description = "Standalone CyberCar automation workspace"
requires-python = ">=3.11"
dependencies = [
  "requests",
  "python-dotenv",
  "websocket-client",
  "pillow"
]

[project.optional-dependencies]
runtime = [
  "DrissionPage",
  "browser-cookie3"
]
dev = [
  "pytest"
]

[project.scripts]
cybercar = "cybercar.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
